why do i keep rewriting shit
