+++ # oops another rewrite/SN_APIC.py
