+++ # oops another rewrite/Trinity.cfg
