+++ # oops another rewrite/Trinity.sh
