+++ .env.example
@@ -1,0 +1,2 @@
+SSL_Cert = /System/Keys/Certificate.pem
+SSL_Key = /System/Keys/Certificate.key
